refactor(crashlytics): address PR review comments for web onboarding
- Move firebasetelemetry client utility from src/gcp/ to src/crashlytics/
- Add updateMask parameter to TelemetryConfig PATCH requests
- Parallelize firebasetelemetry and firebasetelemetryadmin API enablement
- Simplify billing prompt feature name to 'Crashlytics'
- Clarify crashlytics:onboard command description for web apps
- Revert unrelated changes to src/deploy/extensions/validate.ts

TAG=agy
CONV=0cf98957-1b4e-437c-8a79-5ff873b6c315
