+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-- Added `crashlytics:onboard` CLI command to support Crashlytics web onboarding.
+- Added `crashlytics:onboard` CLI command to support Crashlytics onboarding for web apps.
 - Added `firebase ailogic:providers:*` CLI commands to enable, disable, and list Gemini API providers.
 - Add `MCP-Protocol-Version`, `Mcp-Method`, and `Mcp-Name` HTTP headers to `OneMcpServer` requests per the MCP 0728 standard release candidate (https://blog.modelcontextprotocol.io/posts/2026-07-28-release-candidate/ and https://modelcontextprotocol.io/seps/2243-http-standardization).
 - Fixes Storage Emulator to support JSON uploads larger than 100KB without hanging or throwing 413 error (#8355)
